Split out result class
