+++ java/src/main/java/com/example/Result.java
@@ -1,0 +1,13 @@
+package com.example;
+
+class Result
+{
+    Double value;
+    Double stddev;
+
+    public Result(Double value, Double stddev)
+    {
+        this.value = value;
+        this.stddev = stddev;
+    }
+}
